개선할 부분 기록용
status의 초기화까지 FSM에서 처리하다 보니 확실히 남들이 보기엔 구조가 엉망인것처럼 보일 수 있음..
 FSM에서 status를 멤버를 갖도록하는것은 그대로 두되, 속성이 아닌 behaviour로 두어서 컴포넌트로 갖도록 할까 고민중.. 초기화도 start or enable에서 할 수 있으니

+++ 1'st RPG/Assets/Scripts/Player/Status.cs
@@ -13,8 +13,8 @@ namespace common
         int _hp;
         int _level;
 
-        int _exp;
-        int _maxExp;
+        float _exp;
+        float _maxExp;
 
         int _atk;
         float _moveSpeed;
@@ -24,8 +24,8 @@ namespace common
         public int Level { get { return _level; } set { _level = value; } }
         public string Name { get { return _name; } set { _name = value; } }
 
-        public int MaxExp { get { return _maxExp; } set { _maxExp = value; } }
-        public int Exp { get { return _exp; } set { _exp = value; } }
+        public float MaxExp { get { return _maxExp; } set { _maxExp = value; } }
+        public float Exp { get { return _exp; } set { _exp = value; } }
 
         public int Atk { get { return _atk;} set { _atk = value; } }
